refactor(transactions): move components into features/transactions/
Relocates TransactionTable, Filters, TransactionForm, and the three
action files into features/transactions/. Adds a barrel index.ts.
Page files updated to use @/features/transactions/* absolute imports.
Zero behavior change; all 74 tests pass.
